Refine Rust numeric safety and solution boundaries
Teach explicit numeric overflow and conversion policies, fix typed division overflow and the health OpenAPI contract, and remove starter-only error variants from completed references.

Copilot-Session: 3bcd9f90-d99d-43e9-8000-90d3cfe01dcc

+++ lessons/01_basics/03_strings_and_functions.rs
@@ -11,8 +11,9 @@ fn greeting(name: &str) -> String {
     format!("Hello, {name}!")
 }
 
-fn rectangle_area(width: u32, height: u32) -> u32 {
-    width * height // tail expression: the function's return value
+fn rectangle_area(width: u32, height: u32) -> u64 {
+    // Lossless widening makes every possible u32 × u32 product fit in u64.
+    u64::from(width) * u64::from(height)
 }
 
 fn classify_length(text: &str) -> &'static str {
